+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2024-05-16-1
+
+Changes:
+
+- Use a function to export es instead returns
+
 ## v2024-05-14-1
 
 Changes: